Rename date parser interface and class
